accounting/requirements.txt
fastapi[all]
tortoise-orm[aiosqlite]
faust